+++ railties/lib/rails/generators/rails/app/templates/config/application.rb
@@ -9,7 +9,7 @@ require "active_job/railtie"
 <%= comment_if :skip_active_record %>require "active_record/railtie"
 require "action_controller/railtie"
 require "action_mailer/railtie"
-<%= comment_if :skip_action_view %>require "action_view/railtie"
+require "action_view/railtie"
 <%= comment_if :skip_sprockets %>require "sprockets/railtie"
 <%= comment_if :skip_test_unit %>require "rails/test_unit/railtie"
 <% end -%>